Check if `file` command is available at require time.
FileInfo does not have a hard link to the `file` command, which means
that it being absent or removed at any point in time, will break the gem
without notice to anyone using it.

To make FileInfo gem more reliable in a changing environment, we check
for the `file` command at the gem load time, and issue an `LoadError`
if the command happens to be unavailable.